Update Main.java
Took out the first switch and made a few if else if statements because I couldn't figure out the switch. seems to be working now. Also created a new gameMenuPrompt to give the user options to choose when they are playing, and created another switch with the selection of each option.