Update Main.java
shifted boolean exit = false to towards the top of the page to declare it, added in exit = true to selection of option 0, modified the do/while statement after the first do/while so that if option 0 is selected, it will end the program. also added in exit = true to the if / else if statement at the end of Main before the private static void method menuPrompt so that when those requirements are met it exits.