Update EnemyStats
Take a look at the updated file I wrote starting on line 32. I modified it so that it changes the group members that was created in userstats and not your variable in this class. I also modified it so that it takes the parameter of UserStats stats so that it can call stats.decreaseMembers() instead of reducing the variable you made in this class. That way we just have the static variable MAX_MEMBERS in userstats that we are all working with. It is now working in jGrasp. Let me know if you are good with the edits.